pyproject.toml
[project]
name = "lm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
]

[tool.setuptools.packages.find]
include = ["lm*"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "ruff",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]